add methods.py, voice message processing
