+++ constants.py
@@ -1,5 +1,6 @@
 START_MESSAGE = "Привет! Я бот для конвертации голосовых и видео сообщений в текст.  Когда я получу аудио- или видео-сообщение я представлю тебе его в текстовом формате. Также, ты можешь настроить меня, используя ..."
 
-TEMP_AUDIO_FILE_NAME = "voice.ogg"
+TEMP_AUDIO_FILE_NAME = "temp_voice.ogg"
+TEMP_VIDEO_FILE_NAME = "temp_video.mp4"
 
 TOKEN_NOT_SET = "Error: TELEGRAM_BOT_TOKEN environment variable not set."
