+++ requirements.txt
@@ -1,3 +1,4 @@
 pyTelegramBotAPI
 python-dotenv
+moviepy
 git+https://github.com/openai/whisper.git
